More linux compile warning fixes

+++ tensorflow/compiler/plugin/poplar/stream_executor/executor.cc
@@ -367,7 +367,7 @@ PoplarExecutor::ExecuteEngine(poplar::Engine* engine,
       for (size_t a = 0; a < args.size(); a++) {
         auto mem = args[a];
         TensorControl *tc = reinterpret_cast<TensorControl *>(mem.opaque());
-        if (tc->on_device == false || tc->input_handle != a || engine_changed) {
+        if (tc->on_device == false || tc->input_handle != (int64)a || engine_changed) {
           void *buf(static_cast<void *>(tc->data));
           if (input_convertors[a]) {
             std::vector<char> converted = input_convertors[a](buf, tc->size, 0);
